requirements.txt
Django==1.11.4
Pillow
mysqlclient==1.3.12
django-crispy-forms==1.7.0
python-dateutil
django-registration-redux==1.8
social-auth-app-django==2.0.0
django-pagedown==0.1.3
django-modeltranslation==0.18.1
django-markdown-deux==1.0.5
djang-star-rating==0.5.6
django-hitcount==1.2.3
django-admin-tools==0.8.1
django-feedparser==0.2.1
django-simple-captcha==0.5.5